requirements.txt
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.0
pandas>=2.2.0
duckdb>=1.0.0
yfinance>=0.2.40
curl_cffi==0.5.9
pybit>=5.3.0
pyarrow>=15.0.0
schedule>=1.2.1
s3fs>=2024.2.0
requests-cache
requests-ratelimiter
pyrate-limiter